base: move type traits from utils
