+++ src/v/container/tests/vector_bench.cc
@@ -13,7 +13,7 @@
 #include "container/tests/bench_utils.h"
 #include "random/generators.h"
 #include "utils/functional.h"
-#include "utils/type_traits.h"
+#include "base/type_traits.h"
 
 #include <seastar/testing/perf_tests.hh>
